Deleted AdminUser.java
AdminUser will just be a flag instead within the User class. The server can send a request to chat log for the log by passing the current User. The chat log return function can verify that the flag is true, just in case. But the GUI would be setup to only have that option if the current user is an admin, so the chat log check shouldn't really ever return false.